Extended the `register_user` functionality of the `Authenticator` class
Notes:
We need to get the password of the new user so we can store it in the
database. The `register_user` method from `streamlit-authenticator`'s `Authenticator` class does not return the password. We extended the class to make the method return the hased password

References:
Python Class Inheritance Explained | Built In
https://builtin.com/software-engineering-perspectives/python-inheritance
https://builtin.com/software-engineering-perspectives/python-inheritance#:~:text=Class%20Inheritance%20in%20Python&text=The%20class%20that%20has%20the,is%20called%20the%20child%20class.

list - How to flatten a tuple in python - Stack Overflow
https://stackoverflow.com/questions/18500541/how-to-flatten-a-tuple-in-python
https://stackoverflow.com/a/64938679
https://stackoverflow.com/questions/18500541/how-to-flatten-a-tuple-in-python/64938679#64938679

Cannot import name 'hasher' from 'streamlit_authenticator' - 🎈 Using Streamlit - Streamlit
https://discuss.streamlit.io/t/cannot-import-name-hasher-from-streamlit-authenticator/65675
https://discuss.streamlit.io/t/cannot-import-name-hasher-from-streamlit-authenticator/65675/2